+++ app/src/main/java/com/miguelrodriguez19/mindmaster/models/comparators/AccountComparator.kt
@@ -1,6 +1,6 @@
 package com.miguelrodriguez19.mindmaster.models.comparators
 
-import com.miguelrodriguez19.mindmaster.models.structures.GroupPasswordsResponse.Account
+import com.miguelrodriguez19.mindmaster.models.structures.dto.GroupPasswordsResponse.Account
 
 class AccountComparator : Comparator<Account> {
     override fun compare(o1: Account?, o2: Account?): Int {
