+++ app/src/main/java/com/miguelrodriguez19/mindmaster/models/comparators/EventGroupComparator.kt
@@ -1,6 +1,6 @@
 package com.miguelrodriguez19.mindmaster.models.comparators
 
-import com.miguelrodriguez19.mindmaster.models.structures.EventsResponse
+import com.miguelrodriguez19.mindmaster.models.structures.abstractClasses.EventsResponse
 import java.text.SimpleDateFormat
 import java.util.*
 import kotlin.Comparator
