+++ app/src/main/java/com/miguelrodriguez19/mindmaster/models/firebase/FUserManager.kt
@@ -1,11 +1,10 @@
 package com.miguelrodriguez19.mindmaster.models.firebase
 
 import android.content.Context
-import android.media.MediaCodec.MetricsConstants.SECURE
 import com.google.firebase.auth.FirebaseAuthUserCollisionException
 import com.google.firebase.auth.FirebaseUser
 import com.miguelrodriguez19.mindmaster.R
-import com.miguelrodriguez19.mindmaster.models.structures.UserResponse
+import com.miguelrodriguez19.mindmaster.models.structures.dto.UserResponse
 import com.miguelrodriguez19.mindmaster.models.utils.AllDialogs
 import com.miguelrodriguez19.mindmaster.models.utils.Preferences
 import kotlinx.coroutines.Dispatchers
