+++ app/src/main/java/com/miguelrodriguez19/mindmaster/models/formatters/ChartsValueFormatter.kt
@@ -17,12 +17,4 @@ class ChartsValueFormatter : ValueFormatter() {
         }
     }
 
-    override fun getFormattedValue(value: Float): String {
-        return if (value % 1 == 0f) {
-            "${value.toInt()}${Preferences.getCurrency()}"
-        } else {
-            val decimalFormat = DecimalFormat("0.00")
-            "${decimalFormat.format(value)}${Preferences.getCurrency()}"
-        }
-    }
 }
