+++ app/src/main/java/com/miguelrodriguez19/mindmaster/models/structures/dto/GroupPasswordsResponse.kt
@@ -1,4 +1,4 @@
-package com.miguelrodriguez19.mindmaster.models.structures
+package com.miguelrodriguez19.mindmaster.models.structures.dto
 
 import java.io.Serializable
