+++ app/src/main/java/com/miguelrodriguez19/mindmaster/models/structures/enums/EventType.kt
@@ -1,0 +1,5 @@
+package com.miguelrodriguez19.mindmaster.models.structures.enums
+
+enum class EventType {
+    EVENT, REMINDER, TASK
+}
