+++ app/src/main/java/com/miguelrodriguez19/mindmaster/models/structures/enums/Status.kt
@@ -1,0 +1,5 @@
+package com.miguelrodriguez19.mindmaster.models.structures.enums
+
+enum class Status {
+    COMPLETED, PENDING, IN_PROGRESS, CANCELLED
+}
