+++ app/src/main/java/com/miguelrodriguez19/mindmaster/models/viewModels/expenses/ExpensesViewModel.kt
@@ -3,7 +3,7 @@ package com.miguelrodriguez19.mindmaster.models.viewModels.expenses
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.miguelrodriguez19.mindmaster.models.structures.MonthMovementsResponse
+import com.miguelrodriguez19.mindmaster.models.structures.dto.MonthMovementsResponse
 
 class ExpensesViewModel : ViewModel() {
     private val _actualMonth = MutableLiveData<MonthMovementsResponse>()
